requirements/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
twilio==8.10.3
nltk==3.8.1
spacy==3.7.2
python-dotenv==1.0.0
transformers==4.36.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
